Fix a bug when used without any extra manifest
With `-A` flag `ls -A /etc/kubeadm/manifests/extra` yields these:

..2025_05_12_10_47_49.2834907494  ..data

Which made the check pass, but applying the dir resulted in script
getting stuck in retry loop.

Simply removing `-A` is not sufficient as kubectl does pickup dot
files, and we cannot miss them. So the solution is ot look inside
of the `..data` dir.