Persist containerd data and use default path
The whiteout error has resurfaced, but this time in tmpfs and not 9p,
so not clear what is the relea cause of it, but since files need to
be persisted anyhow, moving them to a proper filesystem makes sense.
There is a PV already, so containerd can use it.

See original post about 9p: https://github.com/containerd/containerd/issues/3762#issuecomment-608502595